Speed up loading times slightly
In some scenes, the buffer upload is taking the bulk of the time. It's
still not fast due to excessive syncing, but it's at least slightly
better with the newer VMA version and merging of multiple buffer uploads
to one. But optimally, they should all be in one. Though the
acceleration structure compaction becomes more difficult in that case...