Path space regularization
This one's technically slightly biased (in addition to the normal bias
added by regularization). In spatial reuse, reconnection shift assumes
that the tail of the path is correct as-is, but that's not exactly the
case with regularization, since it would change the roughnesses of
vertices in the tail based on the head.